IPRangeColumn (in progress)
