+++ Arriba/Arriba/Arriba.csproj
@@ -88,6 +88,7 @@
     <Compile Include="Model\Column\ColumnDetails.cs" />
     <Compile Include="Model\Column\FastAddSortedColumn.cs" />
     <Compile Include="Model\Column\IpRangeColumn.cs" />
+    <Compile Include="Structures\IpRange.cs" />
     <Compile Include="Model\Correctors\ColumnSecurityCorrector.cs" />
     <Compile Include="Model\Correctors\JoinCorrector.cs" />
     <Compile Include="Model\Correctors\ColumnAliasCorrector.cs" />
